Split panel section in different files
